Subsume Client into Node
