+++ README.md
@@ -22,47 +22,47 @@ Or install it yourself as:
 
 ## Usage
 
-There are two ways to use this gem.  You can make direct calls to the RPC client using Ruby hashes or you can use proxy objects for terser code.
+There are two ways to use this gem.  You can make direct RPC calls using Ruby hashes or you can use proxy objects for terser code.
 
 ### Raw RPC Calls
 
-The RCP client exposes raw Remote Procedure Call methods according to the [Nano RPC Docs](https://github.com/clemahieu/raiblocks/wiki/RPC-protocol).
+The NanoRpc::Node object exposes raw Remote Procedure Call methods according to the [Nano RPC Docs](https://github.com/clemahieu/raiblocks/wiki/RPC-protocol).
 
 Every method requires an `action`, which is passed as the first argument to `call`.  Depending on the action, there may be additional required or optional parameters that are passed as an options hash.
 
-First setup the client:
+First setup the node connection:
 
 ```ruby
 # Connect to the default node (localhost:7076)
-client = NanoRpc.client
+node = NanoRpc.node
 
 # or connect to a custom node
-client = NanoRpc::Client.new(host: 'mynanonode', port: 1234)
+node = NanoRpc::Node.new(host: 'mynanonode', port: 1234)
 ```
 
 If you're using [Nanode](https://www.nanode.co/) or similar service that requires `Authorization` key in HTTP header, you can specify it using `auth`.
 
 ```ruby
-client = NanoRpc::Client.new(auth: 'someauthkey')
+node = NanoRpc::Node.new(auth: 'someauthkey')
 ```
 
 You can also specify custom headers as a hash. These will be sent with every RPC request.
 
 ```ruby
-client = NanoRpc::Client.new(headers: { 'Authorization' => 'someauthkey' })
+node = NanoRpc::Node.new(headers: { 'Authorization' => 'someauthkey' })
 ```
 
-Once the client is created, make a `call`, passing the action and data:
+Once the node is setup, make a `call`, passing the action and data:
 
 ```ruby
-  client.call(:account_balance, account: 'xrb_someaddress1234')
-  # => {"balance"=>100, "pending"=>0}
+node.call(:account_balance, account: 'xrb_someaddress1234')
+# => {"balance"=>100, "pending"=>0}
 ````
 
 Response data are provided as [Hashie](https://github.com/intridea/hashie) objects with integer coercion, indifferent access, and method access.
 
 ```ruby
-  data = client.call(:account_balance, account: 'xrb_someaddress1234')
+  data = node.call(:account_balance, account: 'xrb_someaddress1234')
   # => {"balance"=>100, "pending"=>0}
   data.balance
   # => 100
@@ -101,11 +101,18 @@ There are also helper methods to bypass repetitive nested calls:
   # => 5
 ```
 
-You can point each proxy object at its own node by passing in a client instance:
+You can point each proxy object at its own node by passing in a node instance:
 
 ```ruby
-  client = NanoRpc::Client.new(host: 'mynanonode', port: 1234)
-  account = NanoRpc::Account.new('xrb_someaddress1234', client: client)
+node = NanoRpc::Node.new(host: 'mynanonode', port: 1234)
+account = NanoRpc::Account.new('xrb_someaddress1234', node: node)
+```
+
+Or, easier, you can access acounts and wallets from the node itself:
+
+```ruby
+account = node.access_account('xrb_someaddress1234')
+wallet = node.access_wallet('3AF91AE')
 ```
 
 For a comprehensive guide, see the [Wiki](https://github.com/jcraigk/ruby_nano_rpc/wiki).
