+++ lib/nano_rpc.rb
@@ -5,11 +5,12 @@ require 'nano_rpc/helpers/account_helper'
 require 'nano_rpc/helpers/accounts_helper'
 require 'nano_rpc/helpers/node_helper'
 require 'nano_rpc/helpers/wallet_helper'
-require 'nano_rpc/client'
-require 'nano_rpc/errors'
 require 'nano_rpc/proxy'
+require 'nano_rpc/node'
+require 'nano_rpc/errors'
 require 'nano_rpc/response'
 require 'nano_rpc/proxies/account'
 require 'nano_rpc/proxies/accounts'
-require 'nano_rpc/proxies/node'
 require 'nano_rpc/proxies/wallet'
+
+require 'pry'
