+++ lib/nano_rpc/version.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 module NanoRpc
-  VERSION = '0.9.0'
+  VERSION = '0.10.0'
 end
